services/payment-service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
motor==3.3.1
pymongo==4.6.1
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
aioboto3==9.3.0
python-multipart==0.0.6
dnspython==2.4.2 